% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_base.R
\name{adjust_orig_vals}
\alias{adjust_orig_vals}
\title{Adjust non-reference data points}
\usage{
adjust_orig_vals(
  fit_object,
  df,
  orig_dorms,
  orig_vals_mean,
  orig_vals_se,
  study_id = NULL,
  data_id = NULL,
  ref_dorms = NULL
)
}
\arguments{
\item{fit_object}{the object returned by a \code{CWModel} function call}

\item{df}{a data frame with alternative (i.e. non-reference) data points to be adjusted}

\item{orig_dorms}{name of the column specifying the definition/method}

\item{orig_vals_mean}{column name for the dependent variable in linear space (i.e. not log or logit space)}

\item{orig_vals_se}{column name for the standard error of the dependent variable in linear space}

\item{study_id}{column name; if specified, predicts out on the random effects estimated by CWModel for groups with the same study_id value}

\item{data_id}{column name with row-specific IDs that will be output with the predictions/adjusted values}
}
\value{
a five-element list; mean and SE of the adjusted values, mean and SE of adjustment factor, and data ID
}
\description{
\code{adjust_orig_vals} adjusts data points that used an alternative (i.e. non-reference)
definition/method, based on the fit of the meta-regression, and leaves alone data points
that used the reference definition/method
}
\examples{
preds1 <- adjust_orig_vals(
  fit_object = fit1,
  df = df_orig1,
  orig_dorms = "obs_method",
  orig_vals_mean = "meanvar",
  orig_vals_se = "sdvar"
)


}
