% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_base.R
\name{CovModel}
\alias{CovModel}
\title{Specify a covariate}
\usage{
CovModel(
  cov_name,
  spline = NULL,
  spline_monotonicity = NULL,
  spline_convexity = NULL,
  soln_name = NULL
)
}
\arguments{
\item{cov_name}{name of the covariate}

\item{spline}{the object returned by an \code{XSpline} function call}

\item{spline_monotonicity}{may be "increasing" or "decreasing"; default NULL}

\item{spline_convexity}{may be "convex" or "concave"; default NULL}

\item{soln_name}{an alternative label for the covariate; default NULL}
}
\description{
\code{CovModel} records information about the functional form of a covariate (e.g. monotonically increasing spline)
}
\examples{
CovModel("x1")


}
