% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_base.R
\name{CWData}
\alias{CWData}
\title{Format the data for meta-regression}
\usage{
CWData(
  df,
  obs,
  obs_se,
  alt_dorms,
  ref_dorms,
  dorm_separator = NULL,
  covs = list(),
  study_id = NULL,
  data_id = NULL,
  add_intercept = TRUE
)
}
\arguments{
\item{df}{a data frame}

\item{obs}{column name for the dependent variable (note: must be difference in logs or difference in logits)}

\item{obs_se}{column name for the standard error of the dependent variable}

\item{alt_dorms}{name of the column specifying the alternative definition or method}

\item{ref_dorms}{name of the column specifying the reference definition or method}

\item{dorm_separator}{string indicating separator for multiple dorms in alt_dorms or ref_dorms}

\item{covs}{list of column names to be used as covariates}

\item{study_id}{name of the column for indicating group, for estimating between-group heterogeneity, optional}
}
\value{
an object of type 'environment'
}
\description{
\code{CWData} prepares the data as an input for \code{CWModel}
}
\examples{
dat1 <- CWData(
  df = mydat,
  obs = "ratio",
  obs_se = "ratio_se",
  alt_dorms = "altvar",
  ref_dorms = "refvar",
  covs = list("x1", "x2")
)


}
