% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_base.R
\name{CWModel}
\alias{CWModel}
\title{Fit a model}
\usage{
CWModel(
  cwdata,
  obs_type,
  cov_models,
  gold_dorm,
  order_prior = NULL,
  use_random_intercept = TRUE,
  gamma_bound = NULL,
  max_iter = 100L,
  inlier_pct = 1,
  outer_max_iter = 100L,
  outer_step_size = 1
)
}
\arguments{
\item{cwdata}{the object returned by a \code{CWData} function call}

\item{obs_type}{"diff_log" or "diff_logit" corresponding to whether the dependent variable is in log or logit space}

\item{cov_models}{a list of \code{CovModel} function calls;
must include an element with \code{CovModel("intercept")} to run a model with an intercept}

\item{order_prior}{an optional list of 2-element vectors where the
definition/method specified in the first element is required to
yield a lower predicted value than the definition/method specified in the second element,
all other things being equal}

\item{use_random_intercept}{If you set use_random_intercept=True in CWModel and NOT pass in study_id , a user warning will be printed out}

\item{gamma_bound}{Bounds for a uniform prior on gamma, e.g. \code(array(c(0.0, 0.4)))}

\item{max_iter}{the maximum number of interations allowed for the model to converge; default 100L;
must be integer, e.g. \code{100L} or \code{as.integer(100)}}

\item{inlier_pct}{a float in [0, 1] indicating the proportion of data points presumed to be inliers}

\item{outer_max_iter}{integer; default 100L}

\item{outer_step_size}{numeric; default 1.0}
}
\value{
an object of type 'environment'
}
\description{
\code{CWModel} fits a meta-regression model
}
\examples{
model1 <- CWModel(
  cwdata = dat1,
  obs_type = "diff_log",
  cov_models = list(CovModel("x1"), CovModel("x2")),
  gold_dorm = "measured"
)


}
