% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{delta_transform}
\alias{delta_transform}
\title{Delta method approximation}
\usage{
delta_transform(mean, sd, transformation)
}
\arguments{
\item{mean}{a vector of means}

\item{sd}{a vector of standard deviations}

\item{transformation}{one of "linear_to_log", "log_to_linear", "linear_to_logit", or "logit_to_linear"}
}
\value{
a data frame with the transformed means and standard deviations
}
\description{
\code{delta_transform} uses the delta method approximation to transform random variables to/from log/logit space
}
\examples{
df <- data.frame(meanvar = c(0.9, 1.0, 1,1), sdvar = (0.2, 0.2, 0.3))
delta_transform(mean = df$meanvar, sd = df$sdvar, transformation = "linear_to_log")
  cwdata = dat1,
  obs_type = "diff_log",
  cov_models = list(CovModel("x1"), CovModel("x2")),
  gold_dorm = "measured"
)


}
