% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_base.R
\name{XSpline}
\alias{XSpline}
\title{Specify a spline}
\usage{
XSpline(knots, degree, l_linear, r_linear)
}
\arguments{
\item{knots}{a vector of knot locations, including knots at <= min and >= max}

\item{degree}{polynomial degree (1=linear, 2=quadratic, 3=cubic); must be integer, e.g. \code{3L} or \code{as.integer(3)}}

\item{l_linear}{boolean indicating whether the leftmost segment is restricted to be linear}

\item{r_linear}{boolean indicating whether the rightmost segment is restricted to be linear}
}
\description{
\code{XSpline} records information about the functional form of a spline (e.g. degree, knot locations)
}
\examples{
XSpline(knots = c(3,4,5,6,7), degree = 3L, l_linear = TRUE, r_linear = TRUE)


}
